pyproject.toml
[project]
name = "jeffs-api-ripper"
version = "0.1.0"
description = "Jeff's API Ripper - Dealer menu API details extractor with Streamlit interface"
authors = [
    {name = "Jeff Jackson", email = "jeff99jackson99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "lxml>=4.9.0",
    "pygithub>=1.59.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "Q", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
